HID: Add list for handling set report from kernel
This patch adds a list to keep track of pending set reports from
kernel, so that once the set report completes, the coresponding report
id can be sent back to uhid driver to unblock which was waiting for
set report to complete. Also some modifictions are done to enhance
the existing logic for making the uhid user spece driver as non blocking.

Change-Id: I86001997656a9e7c15b41a6afda705bff4e117e9
CRs-Fixed: 963213