A2DP audio HAL: fix write error behavior
Do not return -1 status in case of write error but
emulate normal timing and return the number of bytes written
as if no error had occured.

This is because audio flinger does not take any recovery
action in case of write error anyway but does not increment the
number of frames written to the HAL.
In case of persisting error, active audio tracks will not finish playback
and a wakelock will be held indefinitely.

Bug: 25488825
Change-Id: I4f00ba8d7a5c55f70520d448d894240c54685027
(cherry picked from commit ec647fa1db61806a334fc883fb98679768ac4d0b)