Stops all BLE timers when BT turn off request is issued to BT stack
Stops all the BLE timers namely the multi advertisement
RPA timers, the local RPA timer and the BLE GAP timers
as soon as BT turn off request is issued to the BT stack
rather than waiting for the BT turn off callback to stop
the timers. This prevents BLE timers from starting when
the BT turn off request is in process.

CRs-fixed: 974344

Change-Id: I0b47a9ae892b7cb226d2ce7d7e30b41d0ecad582