Bluetooth: Handle connection collision scenario
Root cause: We are getting connection request from the
remote at the same time as we are sending a create
connection command. This leads to the host deleting the
LCB causing the issue.

Fix: Declare a variable in LCB which keeps track of
collision scenario, and in those cases avoid deleting the
LCB from the host side.

CRs-Fixed: 999229
Change-Id: I6f9a7971a3921f01c740d2097bf2a68bbf00a5d6