Don't unset the HF indicators and S4 Settings for PTS
PTS Test Case HFP_AG_SLC_BV_10_I failed due to PTS HFP version is 1.5.
If there is HFP Version less than 1.7 we are unsetting the HF indicators
and S4 settings in BRSF.So,regardless of PTS HFP version we should not
unset the HF indicators and S4 settings in BRSF.
Also for HSP Profile some PTS Test cases failed because when there is
HSP connection we are checking for callsetup indicator which were not
present for HSP profile. So, we should check for HFP connection as well
with the call setup Indicator.

CRs-Fixed: 1000512
Change-Id: I61239a187c2f85b759eb3029c47e6c2fe341c16e