Bluetooth: Fix DUT cannot ccnnect remote headset BT off/ON
When DUT try to connect a paired power-off headset, there is
no ACL connection at all. Then do BT off, the av sm is still
opening,it will not move to idle and clear the connection
request in the btif layer. At the same time if the Adpater
service is not unbinded during BT off, the btif-queue will
not tear down. Next time BT on, the last connection request
is actully invalid but its state is still busy in the
connection list, then it will block the new connection
request from up layer. The fix will set the sm back to idle
and remove the ongoing connnecton request if sm is opening
state before BT off.

CRs-Fixed: 1008846
Change-Id: Id953cb811107d1cc0ccaa3683027d61fdfa90fb9