HID: Fix regression caused by commit "3a2c2d61"
uhid_event function returned 0 in normal case and with below
commit polling thread was being exited resulting in user space
not able to read the events sent by uhid kernel driver leading
to IOT issues with HID devices using the vendor hid usb drivers.
Eg: Magic Mouse scroll functionality was not working as
for normal functionality of Magic mouse, output report
needs to be sent to hid device, which was not being sent with
current implemenation.

commit 3a2c2d61f4c8769ea9c1615208a6daf2fa162084
Author: Myles Watson <mylesgw@google.com>
Date:   Tue Nov 10 14:35:57 2015 -0800

    Retry HID Polling when EINTR is returned

    While I'm here
     - use __func__ consistently in btif_hh_poll_event_thread.
     - remove braces around a single line

    BUG: 24647083
    Change-Id: I80b99ff0d3ba6336f5ab7a517269fdf63df59b37
    Signed-off-by: Myles Watson <mylesgw@google.com>

Change-Id: Ifa819abd8b0939ba8cc23bdda49be2e16deace18