BT: Changes to cancel alarms according to registered callback
These changes are to cancel all alarms that are registered
to call a specific alarm.
This function will be called to cancel all alarms whose
callback is processed by queue that is about to be cleared

Changes done to fix the crashes when alarm expired to call
the callbacks but the processing queue is already cleared
during shutdown.

Change-Id: Icec85814566d0b5ccdea4312d34cada7da4459d5
CRs-Fixed: 1028196