BT Property: Fix to return the proper default value from property_get
There was an issue to have the default value copied incase particular property
is not found in the database. This fixes the proper return of default value passed

Also added a common path to save the output PCM File from the A2DP SINK path.
This path is a common path across platforms.

Change-Id: Ibfb88d98f9f0adcc269f8d1d26db4ee37d1bf5dc