Bluetooth: fix map navigation voice is not clear
when the link condition is not well, the NOC packet will delay for
about 200ms, while the data queue size is 10, and BT host will drop
the audio packet during stuck. This will cause voice not clear, set
the queue value back to 18, it can cache more data packet, and will
send the cached data when link condition recovers.

CRs-Fixed: 1057518
Change-Id: I99076270d2a03e9e35d3d14861606b5247eb3f1a