Only add RPA devices to the resolving list
This change makes sure that the remote devices which are being
added to the resolving list are RPAs and they do not have a public
address type.
This is needed to address the new privacy specification where,
by default, the controller only searches for RPAs in the
advertisements if a corresponding entry is present in the
resolving list.

CRs-Fixed: 1080023
Change-Id: I7a925c9faf8004ce631bc73b5a3ae15af8d42796