Use device privacy mode as default
This change configures device privacy mode for each
remote device added to the resolving list.
Before exercising the privacy command it confirms
whether the set privacy HCI command is supported.
Also, this change reverts the previous change made
which avoids the addition of public address devices
to the resolving list.

Change-Id: I7ddb3781dafa263e21754a4be38e708f8299b1a8