Blueooth: Reset the call_ind to 0 if call is held
When the first call is held, and the second incoming or
outgoing call being active, DUT will not update the indicator
value ciev(1,1) to remote device, as it will conclude that
call_ind keeps the same value and no need to update. The fix
will update the indicator value in this situation.

CRs-Fixed: 2117300
Change-Id: Ibb23402fd09a023c397ae7a70145f303b7230453