BLE: Extend to enable/disable BLE dynamically
Retrieve the property below to check whether BLE is supported or not.
(1) vendor.bt.is_ble_supported
// default Bluetooth adapter

(2) vendor.bt.is_ble_supported1
// new Bluetooth adapter

The default value for the property is true, indicating BLE is supported.

If BLE is not supported,
(1) Return NULL in btif_gatt_get_interface

(2) Not register GATT, SMP in L2CAP

(3) Not register gatt when to enable BT

(4) Not send HCI LE vendor command

Change-Id: Id222e8e52cafa84164e5e893ebfd2172f49fa229