Remove oldest device from storage
In current implementation, although the maximum number of paired device is 100,
the pairing process for the 101th device can still be successful. However, after
the 101th device is paired, the BT re-enable process is failed due to the stack
overflow.

This code change is to support the use case: If we have exceeded the maximum
number of allowable records to allocate, not only deleting the oldest record to
make room for the new record, but also removing the link information from the
storage.

Change-Id: I3544a1b1af60f1a527f3b13406c6553ca5952476
Signed-off-by: Song Gao <songg@codeaurora.org>