BLE: Devices are not found in the inquiry list
If BLE advt is received before BR/EDR advt and the device is not
discoverable, the device is getting added to bdaddr database
but not to inquiry database. This prevents device getting added to
inquiry database when BR/EDR advt packet is handled. Skip adding the
device to bdaddr database too when device is not discoverable.

Change-Id: I0208c52b4ab5dd4226322405bec1b6a01c98cc4d