Bluetooth: Fix the Audio glitches issue for A2DP SRC and HOGP concurrency
while sending HID RAW DATA call back to app layer for
HID reports and playing A2DP media. BTA thread is blocked.
While sending the HID RAW DATA call back to app HID is polling
for ready data , and it makes BTA thread busy. Due to that
not able to send the media packets to L2CAP.

HOGP: Removed polling for HID data sending to HID driver
in kernel, because we are not using HID driver in LE code.

A2DP SRC : If tx_audio_queue is full , in that scenario it
tries to flush the packets and dequeue the previous packets.
Instead of dropping previous packets. if tx_audio_queue is full
we are dropping the current media packets.

Change-Id: I72697286a35d0b0d1e070c15f4016afbc73866a8