Revert "HF_Client: Fix SCO connection issue"
Description:
This reverts commit 67e8d7709444aa70be6f312cfdf8447e08678303 which will
lead to wrongly selecting of esco parameters during creating eSCO/SCO.

Root cause:
With this change, the first sco_db will always be selected regardless
of the AG the call is for. So default esco paramter will be selected
if HF call is for second AG as default codec is selected. This is not
correct if WBS is supported as new codec shall be selected through
codec negotiation before creating eSCO/SCO connection(Remember we have
done codec negotiation for second HF connection only, not for first one).

Change-Id: Id1be0b89e5e8f255872140cf4e8ef3c5ab0db93c