Reconnect: Workaround to reconnect with advertisers that advertise with public address in the advertisement and do not have any IRK info.
For background connections, we use own_address_type as random identity
to enable resolving and for privacy. This forbids re-connecting with the
advertising packets which have intiator address type as public. We also
skip adding them to resolving list as they do not have capability to decode
address from random address in the connection indication packet.

Change-Id: Id6d3926cba8e8486ed6802b1ef7f8c2f2899a645