AV: cancel the both av and avk accept signalling timer.
In the bta ar module, the AVDT_CONNECT_IND event is propagatd to both AV and AVK when
the remote side supports both A2dp Source and Sink, and then the AV and AVK accept
Signalling timers are created, but only one would be cancelled in time. When the other
one timeout, it would lead to the termination of the acl link.
To fix it: 1, the AVDT_SETCONFIG_CMD_EVT is removed, this can avoid the SEP type lost;
2, cancel both the accept signaling timer when set_config event is handled.

Change-Id: I24d49a133d2e59d3e7bbbda11a897d4b0312b37f
CRs-Fixed: 2478390