Pin key missing: Send callback to upper layers with SMP_CONFIRM_VALUE_ERROR
Some devices reply with pin key missing status and are stuck in that state
when kept in pairing mode. Also, DUT might not have UI to modify state of such devices.
So, we send a callback with SMP_CONFIRM_VALUE_ERR when we get pin key missin error from
such devices. Also, in stack, while calculating confirm value, local address is random.
But, for connection indication packet in bg connection we use public address as
initiator's address for some devices. Such devices calculate with remote address as public
and we calculate as random which results in authentication failure. For such devices, we use
initiator's address as public with blacklisting method.

Change-Id: Ib01bbf2da61a983536c1b3138efb4bd7a338b586