Device Info: Checking for valid pseudo address in security device record
A security device record is allocated when a connection request is being
sent. However, if connection request is not successful, all the entries in
it are not complete. While reading Device info, there are conditional checks
on psuedo address, however verification if pseudo address is valid or not is
not done. So, the condition checks fail and Device info returned is incorrect.

Change-Id: I784d1e2d0c13f0847247b9b8e745341ff3ff2927