A2dpSrc: Handle AVDTP_DISCOVER fail rsp while incoming.
Some remotes which are SBC only supported devices, are not able to
handle START Req, due to codec setup for the peer has been failed
due to remote is not able to handle AVDTP_DISCOVER command while
incoming connection. Due to this DUT is dropping the AVDTP connection
resulting in no audio issue.

- This fix would take care, handling of AVDTP_DISCOVER negative rsp
  from remote and would do get_caps for the same sep while remote does
  set_config.

- AAC frame control fetching is modified according to 5.0 implementation.

CRs-Fixed: 2494211
Change-Id: Icd4711d8ea3de83e2a98d4d5089f12e5d0ad48f1