BT: Clear the command queue while shutting down hci module
In BT on/off scenarios few times while turning off the
BT, few commands are still queued in command queue to send
to soc. but hci module shut down is completed and command queue
is not freed. So in next BT ON scenario, as soon as hci module
is started. It start pumping the pending commands in command
queue. Due to that seeing LE connection complet event for
cancel connection before DEVICE UP state.
To avoid this, if any commands still pending in command queue
need to clear the command queue in HCI module shutdown.