Bluetooth: Disallow to read BD address if BT isn't turned on
Local device tries to read BDADDR as soon as BT is turning on.
"Controller not Ready" error occurs because BT enable flow is not done yet.

BT stack will get adapter's properties including BDADDR once BT enabled,
and callback them to the upper layer.

So there is no need to get BDADDR as soon as BT is turning on.

Change-Id: If9cc53cb2f32a0d12a9992bad9b8c58d92fbf6e8