BT: L2CAP ccb, lcb memory free in BT Disable
Issue: In few scenarios stack is receiving the connect
request and BT disable at same time. Due to that
L2CAP connection is not completed. So that whatever
memory is allocated in connection sequence is not
freed in Disable sequence.
To fix it, in disable sequence deleting the memory
for l2cap ccb, lcb data.

Change-Id: I98de628149c6112e17de3048bc92b41684585237