AVRCP_CT: Fix transaction cannot be released
Issue:
1. Connect with AVRCP TG 1 and AVRCP TG 2
2. Disconnect with TG 1
3. Play music on TG 1
4. Connect with TG 1
5. Loop 2 and 3 several times
Play status or track information is not updated for TG 1

Root cause:
Every register notification command takes 1 transaction, but the transaction
is not released when receive interim response. For one AVRCP connection, all
transactions are released after disconnected, the issue is not hit. For dual
AVRCP connections, all of the transactions are not released when only one
connection disconneted, and the transactions will exhaust when send more and
more register notification commands.

Fix:
Increase maxium transaction number from 16 to 32 for dual AVRCP connection.
Release transaction when receive interim response

Change-Id: I14dc553400f6bb3caa128929bfbd1a2dc912698a