AVRCP_CT: Clear tag rc_supported_play_pos_changed when AVRCP is disconnected
Description:
Sometimes, Device under test(DUT), known as AVRCP controller, fails
to sync the current playback position when user manually chanage it
in AVRCP target(handset).

Reason:
rc_supported_play_pos_changed is set to true when DUT connects with a
handset which supports EVENT_PLAYBACK_POS_CHANGED. And this tag is not
cleared after AVRCP is disconnected. So it will be true still when DUT
connects with a handset which does not support EVENT_PLAYBACK_POS_CHANGED
As a result, DUT will not start a timer to get playback state of remote
handset periodically.

Solution：
Clear tag rc_supported_play_pos_changed when AVRCP is disconnected

Change-Id: Ib7cf3f80b659cf279f78333929d51873fea60188