Fix crash issue due to PIN code buffer overwritten
For the PTS case GATT/SR/GAD/BV-05-C, BT will crash if we input PIN code with 4 digital
bytes in pairing dialog UI. The code logic is we assume PIN code size shall be 6 bytes
if passkey is used, but not check if the input length is 6 or not. So, when 4 digital
bytes are input, we will access the PIN code  buffer with length 6 instead of 4.It will
cause buffer overwritten.

The change will validate the PIN code length.  If the length is not 6, failure shall be
returned.

Change-Id: I1dc88de21747ff957af19e3f3059e3a83dd148b5
CRs-Fixed: 2796212