Alarm: Stricter condition for judging canceled alarm
In current implementation, a timer is set in AVRCP task for getting
play status periodically from peer device. However, in the use case
of turning Bluetooth off, the timer will be canceled in another thread
while AVRCP task is still trying to access the resource of the timer.
This might result in null pointer dereference exception at run time.
This code change is to provide a stricter judging condition which can
prevent the exception.

Change-Id: Ie29e242fc1b08e2b93a9ee51ff25a625fd7ca4e3