HFP_HF/A2DP_SNK: Fix auto connection takes too much time after Bluetooth restart
Use Case:
1. Connect HFP/A2DP with remote device
2. Crash Bluetooth process(com.android.bluetooth)
3. Check if profiles can be connected after Bluetooth is re-enabled

Issue:
It takes more than 1 minute to connect all the required profiles

Reason:
After Bluetooth process is crashed, the remote device is still in
connected state untill link supervision timeout.
Thus the ACL connection request sent from DUT is rejected with reason
code CONNECTION ALREADY EXISTS (0x0B).
Meanwhile, a timer with timeout value set to 60 seconds is initated.
All the retries of profile connection are ignored before the timer expires
This is because the invalid ACL connection is not cleared.

Solution:
1. Reduce L2CAP_LINK_CONNECT_TIMEOUT_MS from 60 seconds to 10 seconds
The is becasue both the typical supervision timeout and page timeout are smaller
than 10 seconds.
2. Set BTA_HF_CLIENT_ACL_COLLISION_TIMER_MS to L2CAP_LINK_CONNECT_TIMEOUT_MS
This is to avoid retrying HFP connection on an invalid ACL connection.

Note:
No change is required for profile A2DP because retry of A2DP connection is based
on its stream statemachine in BTA layer.

Change-Id: I1fa2070330b3079f883af33c1058745ee256532b