A2DP: Set direction for flow specification according to A2DP role
The Flow_Specification command is used to specify the flow parameters for the
traffic carried over the ACL connection, and the flow parameters refer to the
outgoing or incoming traffic of the ACL link, as indicated by the
Flow_direction field.

For A2DP source, set flow direction to outgoing to indicate traffic sent
over the ACL connection.

For A2DP sink, set flow direction to incoming to indicate traffic received
over the ACL connection.

Change-Id: Ia7fa1eec7efe135cd010660e66bc3b90574b0c0a