AVRCP_TG: Enable/Disable absolute volume per property setting
Absolute volume for AVRCP target is disabled when the property
persist.bluetooth.disableabsvol is set to true. Otherwise, absolute
volume is enabled.

In detail, this change is to add/remove AVRCP target to enable/disable
absolute volume for AVRCP target.
(This scenario is typically found in carkit)

Note:
Default value of this property is empty and is treated as false.

Change-Id: I0ac76adcda17d66245d65515f461b82cccfece26