AVRCP_CT: Restart stack when all the transactions are occupied
There is a chance that transactions are used up and can't be
released in time because Bluetooth stack runs into stuck state.

This change is to forcibly restart stack when above error occurs.

Note:
This change only works in above extreme case and brings no side
effect for normal use case.

Change-Id: I7a66f182b731130dbe4423bbae9bdd634d8daa1d