Fix selinux permission issue
After discussions, we found the right home for this property so that
it can be passed in directly in QEMU emulator command line, making
reboot no longer required and also no need to "-selinux permissive"
anymore.

Adding myself as an owner of this particular branch.

Bug: 180647726
Bug: 182795831
Ignore-AOSP-First: the feature is released in rvc-car-dev.
Test: usb passthru
Change-Id: I170de8d77e6b8abdfb96fc1fb7896689e5323c26