GAP : DUMO (RPA + ADV Audio) forget device fix
Issue:
when DUMO (RPA + ADV audio) device initially connected
with pseudo random address and later Identity address
using the link of pseudo random address. In this
scenario if users forgets the device, it only removing
identity address entry in the bt_config file, not
removing pseudo address entry.

Fix:
If it is ADV audio device, checking it is DUMO or not.
If its dumo, removing the entry of pseudo address also

Change-Id: Ic1dacf71110edfe47abd29092659d6a13d4aa59e