HF_CLIENT: Send connection status callback to uplayer
Issue found:
When the allocation of hf client control block is failed, the
hf connection status in application is blocked in connecting status.

Root cause:
In current implementation of connect_int() in hf client, when hf
client cb is not allocated successfully, it just returns value
BT_STATUS_BUSY without any notification of status changed to uplayer.

Resolution:
Calling callbacks to notify the uplayer that the hf client connection
status should be changed to disconnected.

Change-Id: Id49570f0c5306f1840bf9aed3f0e2652f65deea9