Packet Fragmentation: Always reset fragmentation flag  when layer specific is zero
In some cases, where congestion occurs packets are not being transmitted, after the congestion is cleared.
This is because, partial_segment_being_sent is reset only when layer_specific flag is made 0 by decrementing it.
But layer_specific flag is updated based on NOCP,so it can be directly updated to 0.
Made changes to always send MSG_HC_TO_STACK_L2C_SEG_XMIT event when layer_specific is 0

Change-Id: Ie052c7b28530dc4b6a758642dad9ce8b530fc1a1