BTM: Don't find device with invalid handle
Issue found:
Supposing there are two bonded devices A and B in the paired list
in DUT. Device A removes DUT from its own paired list. User operates
on DUT to initiate connection with device A, authentication procedure
is triggered as expection, and user cancels the request in Bluetooth
pairing code confirmation dialog. The observed result is that both
devices are removed from paired list.

Root cause:
After the first bonding request is canceled, BT stack initiates to
create a new ACL connection for the next profile connection request
in the btif queue. Meanwhile, the external app calls createBond()
to bond with device A. Since the connection handle is not generated
at this moment, the authentication request is rejected by the controller.
btm_sec_auth_complete() wrongly find the device in record with invalid
handle. Since device B is the first device with invalid handle in the
record, it is removed from paired list by mistake.

Resolution:
Refine the implementation in btm_sec_auth_complete(). If the input handle
is a invalid one, calling btm_sec_find_dev_by_sec_state() to acquire the
correct device record.

Change-Id: Idfabf39294cd0f68c58f2c93c68468a933227e98