HF_CLIENT: Add checking before initiating the connection
Issue found:
Limit the maximum HFP connection to 2 on DUT side, two devices(A and
B) are bonded with the DUT. Try to connect the HFP from both DUT and
device A simultaneously. Then, initiate the HFP connection from
device B. The HFP connection may be failed with device B.

Root cause:
Two control blocks on btif are allocated to the same BD address(device A).
The first one is allocated due to incoming HFP connection from deivce A,
the second one is allocated when initiating the outgoing connection to
device A without checking for the same BD address. Then the next HFP
connection from device B failed due to no enough resource on btif.

Resolution:
Check if the HFP connection to the same BD address is already on-going
before initiating the connection.

Change-Id: Id56df124335af89ebfc587bcb90a17cbfb53a7f9