BTA: validate array index of stream control block
(1) Issue
When processing AVRCP discovery event, out-of-bounds access to array
"p_cb->p_scb" and "p_cb->lcb" may occur, which causes the Bluetooth
process to raise an exception.

(2) Root Cause
The indexes number of this two arrays obtained from control block
"bta_av_cb" which is initialized with the value of the advanced A/V
event in the method "bta_av_rc_opened". Incorrect values may cause
the index of the array to be equal to -1.

(3) Resolution
Verify that the index is non-negative to prevent out-of-bounds access
to control block array in function "bta_av_rc_disc_done".

Change-Id: I6fa07e81a3776e7a61042af36c015a29321ae958