A2DP_SNK: Change BtifAvStateMachine to idle when signal channel is down
Issue found:
The status of A2DP connection could be different between DUT and peer device
if user iniates A2DP connection/disconnection on peer device.

Root Cause:
The assignment of callback function p_cback in avdt_ccb_ll_closed() depends
on whether the initiator of A2DP connection is DUT or peer device. If the
initiator is DUT, after the callback function bta_av_conn_cback() completes
the process of SIG_CHG, there is no chance for it to send event "BTA_AV_CLOSE_EVT"
for the second time due to signal channel is disconnected. Therefore, the state of
BtifAvStateMachine is stay in opening state and could not be changed to idle state.

Resolution:
Do not clear the connect mask of the AV control block untill the event
"BTA_AV_CLOSE_EVT" is sent and handled.

Change-Id: I3bec3ad9dc328ac8fca8a2a10676b90115c9d78f