BT_SOCK: Only handle the exact number of events
Issue found:
Initiate profile connections with peer device, assertion can be
observed in bt_scok and MAP connection does not get establishment.

Root cause:
In current implementation of sock_poll_thread(), pfds is partially
cleaned up and initialized according to the totoal number of data
fd registered to be polled. So pfds may retain some dirty events from
data fd which are not valid any longer. Due to the implementation
flaw in process_data_sock(), the potential problem is that the actual
number of fd events processed exceeds the number of fd events that
should be processed.

Resolution:
After a fd event is processed, the value of count must be decremented
and be checked whether it equals zero or not. Zero means that all valid
fd events have been processed.

Change-Id: Ibedd3c1de8f35c1d01dd267e9dbd1aa1a3c53311