A2DP_SNK: Send delay report in main thread to avoid crash
In some case, delay reoport is sent in alarm_default_ca
thread while others are sent in bt_main_thread, if both
has data to send, they will check xmit_hold_q queue of
l2cap and try to send at the same time, and one thead may
get a NULL pointer when peeking queue if the other has sent
it out. The change is to send all data in main thread to
avoid race condition and to check the return value and stop
sending the data if it has been sent out.

Change-Id: I1c95f580167fb8f4a3d3d6cbe1292c2f7139ed36