BT: HFP_CLIENT: add RFC_SRV_CLOSE_EVT event
In bta_hf_client_mgmt_cback, if RFC server is closed, event
header is not updated and sent to bta. Due to this BTA state is
changing to garbage value and resulting in BT crash.

To fix this send RFC_SRV_CLOSE_EVT, when RFC server
is closed and client_cb is NULL.

Change-Id: Iaf6030e5ad03f6fcb522cae3f33ad931bce0b64d