BTA_DM: Check BD address in remote name completion
Issue found:
Occasionally, there is no service UUIDs updated after the
pairing completed.

Root cause:
In current implementation of the function bta_dm_remname_cback(),
when the remote name request is completed, it does not check the
Bluetooth device address of the sender. This may run the risk in
the scenario that when the Host initiates services searching with
multi peer devices in sequence, and one of the remote name request
completion is received in delay for some unknown reason. It makes
the state machine switch to idle state, which blocks the processing
of services searching ongoing.

Resolution:
Add address check in bta_dm_remname_cback().

Change-Id: I2e2cf607f0f7a9557b93abd3919b5f50a07da6bc