BTM_SEC: Set encryption after role changed
In current implementation, after authentication is completed,
encryption will be set immediately when DUT is in master role.
However, in some use cases, just before DUT starts authentication,
it sends command to switch role with target device. If the receiving
of role changed event is not prior to authentication completion,
it misses the chance to set encryption just after authentication.
This might result in the bonding state blocked due to SMP pairing.
The resolution is to add a function btm_sec_role_changed() to set
encryption after role changed when conditions meet.

Change-Id: I4b024de2dca11fe15c193665ef8b4613f11a9b2f