Fix dealloc of EATT channel during failures
When EATT channel creation fails due to insufficient
authentication error from remote device, the pending
indication queue is deallocated for Attribute channel
instead of the EATT channel(which was getting created).

This change updates status of LE connection to
"EATT not supported" after deallocating EATT channel
resources (including the pending indication queue)
which were allocated while the EATT channel was in the
process of getting created.
Hence, the pending indication queue of the concerned
EATT channel is deallocated properly instead of the
Attribute channel.

CRs-Fixed: 3239723
Change-Id: I8e9b1b228bd3c5d9ad676454152407abba21b46a