Add mocking support for now function in AttributionProcessor
The triggering of the code we want to test depends on  the return
value of std::chrono::system_clock::now(). To facilicate testing,
in this patch we add a now_func_ field in
AttributionProcessor and make it call it instead of
std::chrono::system_clock::now(). Mocking `now` is made possible
by passing a custom function to the constructor of AttributionProcessor.

Note:
1. This is a manual cherrypick of I7dd3a0e665f72c27e4d1844f45ec15a8dd1ddb53

Bug: 254774758
Test: refactoring, existing tests still pass
Ignore-AOSP-First: security

Change-Id: I1526d794ddd86f53a189f3ff226bddbff7a487e5
(cherry picked from commit 7378585e68bdccd53e2b713d474229310f0672d0)