AVRCP_CT: Handle error response of GetItemAttribute
Issue found:
Metadata is not updated when media player of peer device is
changed from the native one to a third-party one. This issue
can be reproduced with certain brands of mobile phone.

Root cause:
The status of GetItemAttribute Response is "Does not Exist"(0x09)
and the parameter length is 1. However, in function
"handle_get_metadata_attr_response()", track change callback is
called only when status is "Operation completed without error"(0x04).
Therefore, lack of handing this error scenario results in the display
issue in the media player.

Resolution:
When this issue occur, set status as AVRC_STS_NOT_EXIST(0x00) with
dummy data of metadata and notifies the upper layer through callback.

Change-Id: If7022adcdfa54118a2ccae45cbbb14a0fc992e5c