DM: Don't update empty custom UUID's in EIR response
Issue found:
For certain certification, DM1/DM3 packets are recommended for EIR
packet. However, it is observed that DM5 packet is used if empty UUIDs
are updated.

Root cause:
In current implementation of bta_dm_set_eir(), it does not check
whether the custom UUIDs are empty or not before updating in the EIR.
This results in including redundant data in packet and thus enlarge the
size of the packet.

Resolution:
Always checking the validity of the custom UUIDs before updating the EIR.

CRs-Fixed: 3428630
Change-Id: I66d7e42e6fb0222d9954f7e2be0add7cc85912fc
(cherry picked from commit 34e874749515f75fdcf471cf275ef748fae9cd0f)