BT: Handle SSR for full dump scenario
Issue:In fulldump scenario, glink channels are closed and
BT wont get the dump. However SSR process checks for dumps
to be collected. So BT reset not happening.
Fix: we check for dump and if dump not started in 7 secs,
gracefully close the BT and start the BT RESET procedure.

Change-Id: I8d9c112bd07e6d64281635c2d3eaac859b5e8c1e