BT_A2DP_FIX:Playback is not routing to usbhs
--Observed the issue while inserting a usb hs during ongoing playback
the audio not routing to hs and after removing it's not routing back to
speaker.
  We observed that when new device is attached AAudio stream will close
and we need to restart it again but we are not handing this scenario.
Now we added a code changes to handle this cases when usb hs insert or
removed now we are closing and starting new AAudio stream

CRs-Fixed:3808344, 3808335
Change-Id: Ic4ad3d57b734417f5165ce376fcb10f740cec684
Signed-off-by: Tyagaraju Gowda <quic_tgowda@quicinc.com>