Adding support for multiple LE CoC Connections on same PSM.
When device is listening on a PSM, it is possible for more than
one remote device to initiate socket connection on the same PSM.

A socket pair is created when server starts listening for the socket connections(listen fd, listen app_fd).
When a remote device initiates socket connection, server accepts
it, new socket pair is created when socket connection is accepted on server(accpet fd, accept app_fd), this shall be used for further socket communication.

Added Proto message to communicate both listen fd and accept fd
to slate, so that it can use accept fd for all further transactions
like socket read/write/disconnect, and listen fd is free to listen to
the next incoming socket connection.

Change-Id: I91f1b6ebd08717478ea94d2ea642b7eaeaa1d9ef