CSIP: Add Support for AD Type Filtering for Latest SOC Versions
Issue: The second earbud is not pairing during CSIP device pairing.

RCA: This issue occurs on devices running Android version 14 with
the latest SOC version. The latest SOCs do not support group
filtering, resulting in the second device not being discovered
when the stack attempts to apply group filtering and waits for the second set member to be discovered.

Fix: For devices with the latest SOC version, we are setting
AD type filtering.

CRs-Fixed: 4101704
Change-Id: Ic73b72d86264a9a497a09fe0e0e288882b01cb57