HFP_CLIENT: Removing hfp_client connect request
-- corrected removing of hfp_client connect request from btif connect
queue. We observed that we are trying to remove hfp_client connect
request with invalid bd address during connect failed, because of this
hfp_client connect request remain in the queue and not allowing other
device to connect it. Now we added changes to clear connect request
during connect failed.

Change-Id: I5750f68e95b0907b01ba0a0096bb6957febd4c0d